// packages/frontend/e2e-smoke-tests/interfaces/app-interfaces.ts
export enum EAppUrl {
    dashboard = 'dashboard.html',
    backtestingManager = 'backtesting.html',
    herodotusTerminal = 'herodotus-terminal.html',
    tradingServersManager = 'trading-servers-manager.html',
    tradingStats = 'trading-stats.html',
    balanceMonitor = 'balance-monitor.html',
    wsQueryTerminal = 'ws-query-terminal.html',
}